requirements.txt
flask==1.1.2
flask_restx
requests
python-dotenv
redis
PyGithub